Update README and run script to work again
Well, 'work' after a fashion. bundler is shoddy (see README) and
this version of jekyll is using deprecated features in ruby.

I may look to update the jekyll version in future, but ideally would
shift to something that isn't ruby because this just seems to cause
problems every now and then.