Change 'CONTACT' to 'Contact Manchester'
While the theatre's branding is the word "CONTACT" in capitals, they
often go by the more descriptive "Contact Theatre" (e.g. in the
screen reader-focused h1 heading on their home page, or the title of
their YouTube channel).

I've made this change not to be a pedant, but because I think this
title/link is much more explanatory in the context of the PyCon site.

"Contact Theatre" is also how the venue is identified on Google Maps, so
if we use that name I expect it'll help people find it when they arrive.