Add inline menu at top of /local-info
I've decreased the heading levels as I'm about to start adding more
"top-level" sections to this page (e.g. "Food and drink").

The way the site template is currently written the page's main heading
is marked up as an h2 tag. As the "Getting around" and "Food and drink"
sections are hierarchically beneath the h2, I've demoted all existing
headings by one level.

I think it'd be better to follow the convention of using h1 for the main
heading (and removing h1 for the "logo" in the main header). But that's
a site-wide change.